+++ Engine/Source/Programs/UnrealCloudDDC/Jupiter/Implementation/Blob/AzureBlobStore.cs
@@ -6,7 +6,6 @@ using System.Globalization;
 using System.IO;
 using System.Threading.Tasks;
 using Azure;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
@@ -15,46 +14,59 @@ using Jupiter.Common.Implementation;
 using System.Threading;
 using System.Runtime.CompilerServices;
 using System.Collections.Concurrent;
+using Jupiter.Common;
 using Microsoft.Extensions.Logging;
 
 namespace Jupiter.Implementation
 {
     public class AzureBlobStore : IBlobStore
     {
+        private readonly IOptionsMonitor<AzureSettings> _settings;
+        private readonly ISecretResolver _secretResolver;
         private readonly ILogger _logger;
-        private readonly string _connectionString;
+        private readonly INamespacePolicyResolver _namespacePolicyResolver;
         private readonly ConcurrentDictionary<NamespaceId, IStorageBackend> _backends = new ConcurrentDictionary<NamespaceId, IStorageBackend>();
 
-        public AzureBlobStore(IOptionsMonitor<AzureSettings> settings, IServiceProvider provider, ILogger<AzureBlobStore> logger)
+        public AzureBlobStore(IOptionsMonitor<AzureSettings> settings, ISecretResolver secretResolver, ILogger<AzureBlobStore> logger, INamespacePolicyResolver namespacePolicyResolver)
         {
-            _connectionString = GetConnectionString(settings.CurrentValue, provider);
+            _settings = settings;
+            _secretResolver = secretResolver;
             _logger = logger;
+            _namespacePolicyResolver = namespacePolicyResolver;
         }
 
         private IStorageBackend GetBackend(NamespaceId ns)
         {
-            return _backends.GetOrAdd(ns, x => new AzureStorageBackend(_connectionString, ns, SanitizeNamespace(ns), _logger));
+            return _backends.GetOrAdd(ns, x => new AzureStorageBackend(GetConnectionString(ns), ns, GetContainerName(ns), _logger));
+        }
+
+        private string GetContainerName(NamespaceId ns)
+        {
+            NamespacePolicy policy = _namespacePolicyResolver.GetPoliciesForNs(ns);
+
+            string containerName = !string.IsNullOrEmpty(policy.StoragePool) ? $"jupiter-{policy.StoragePool}" : "jupiter";
+            if (_settings.CurrentValue.StoragePoolContainerOverride.TryGetValue(policy.StoragePool, out string? overriddenContainerName))
+            {
+                containerName = overriddenContainerName;
+            }
+
+            return SanitizeContainerName(containerName);
+        }
+
+        private string GetConnectionString(NamespaceId ns)
+        {
+            NamespacePolicy policy = _namespacePolicyResolver.GetPoliciesForNs(ns);
+            string connectionString = _settings.CurrentValue.ConnectionString;
+            if (_settings.CurrentValue.StoragePoolConnectionStrings.TryGetValue(policy.StoragePool, out string? connectionStringForPool))
+            {
+                connectionString = connectionStringForPool;
+            }
+
+            string resolvedConnectionString = _secretResolver.Resolve(connectionString);
+            return resolvedConnectionString;
         }
 
         private static string GetPath(BlobIdentifier blobIdentifier) => blobIdentifier.ToString();
-
-        /// <summary>
-        /// Gets the connection string for Azure storage.
-        /// If a key vault secret is used, the value is cached in <see cref="AzureSettings.ConnectionString"/>
-        /// for next time.
-        /// </summary>
-        /// <param name="settings"></param>
-        /// <param name="provider"></param>
-        /// <returns></returns>
-        public static string GetConnectionString(AzureSettings settings, IServiceProvider provider)
-        {
-            // Cache the connection string in the settings for next time.
-            ISecretResolver secretResolver = provider.GetService<ISecretResolver>()!;
-            string connectionString = secretResolver.Resolve(settings.ConnectionString)!;
-            settings.ConnectionString = connectionString;
-
-            return connectionString;
-        }
 
         public async Task<BlobIdentifier> PutObject(NamespaceId ns, ReadOnlyMemory<byte> content, BlobIdentifier blobIdentifier)
         {
@@ -80,9 +92,9 @@ namespace Jupiter.Implementation
         {
             BlobContents? contents = await GetBackend(ns).TryReadAsync(GetPath(blobIdentifier), flags, CancellationToken.None);
             if (contents == null)
-                {
-                    throw new BlobNotFoundException(ns, blobIdentifier);
-                }
+            {
+                throw new BlobNotFoundException(ns, blobIdentifier);
+            }
             return contents;
         }
 
@@ -98,8 +110,8 @@ namespace Jupiter.Implementation
 
         public async Task DeleteNamespace(NamespaceId ns)
         {
-            string fixedNamespace = SanitizeNamespace(ns);
-            BlobContainerClient container = new BlobContainerClient(_connectionString, fixedNamespace);
+            string fixedNamespace = GetContainerName(ns);
+            BlobContainerClient container = new BlobContainerClient(GetConnectionString(ns), fixedNamespace);
             if (await container.ExistsAsync())
             {
                 // we can only delete it if the container exists
@@ -115,60 +127,38 @@ namespace Jupiter.Implementation
             }
         }
 
-        private static string SanitizeNamespace(NamespaceId ns)
+        private static string SanitizeContainerName(string containerName)
         {
-            return ns.ToString().Replace(".", "-", StringComparison.OrdinalIgnoreCase).Replace("_", "-", StringComparison.OrdinalIgnoreCase).ToLower();
+            return containerName.Replace(".", "-", StringComparison.OrdinalIgnoreCase).Replace("_", "-", StringComparison.OrdinalIgnoreCase).ToLower();
         }
     }
 
     public class AzureStorageBackend : IStorageBackend
     {
-        private readonly string _connectionString;
         private readonly NamespaceId _namespaceId;
-        private readonly string _containerName;
         private readonly ILogger _logger;
+        private readonly BlobContainerClient _blobContainer;
 
         private const string LastTouchedKey = "Io_LastTouched";
         private const string NamespaceKey = "Io_Namespace";
 
         public AzureStorageBackend(string connectionString, NamespaceId namespaceId, string containerName, ILogger logger)
         {
-            _connectionString = connectionString;
             _namespaceId = namespaceId;
-            _containerName = containerName;
             _logger = logger;
-        }
-
-        /// <summary>
-        /// Gets the connection string for Azure storage.
-        /// If a key vault secret is used, the value is cached in <see cref="AzureSettings.ConnectionString"/>
-        /// for next time.
-        /// </summary>
-        /// <param name="settings"></param>
-        /// <param name="provider"></param>
-        /// <returns></returns>
-        public static string GetConnectionString(AzureSettings settings, IServiceProvider provider)
-        {
-            // Cache the connection string in the settings for next time.
-            ISecretResolver secretResolver = provider.GetService<ISecretResolver>()!;
-            string connectionString = secretResolver.Resolve(settings.ConnectionString)!;
-            settings.ConnectionString = connectionString;
-
-            return connectionString;
+            _blobContainer = new BlobContainerClient(connectionString, containerName);
         }
 
         public async Task WriteAsync(string path, Stream content, CancellationToken cancellationToken)
         {
-            _logger.LogDebug("Checking if Azure container with name {Name} exists", _containerName);
-            BlobContainerClient container = new BlobContainerClient(_connectionString, _containerName);
             Dictionary<string, string> metadata = new Dictionary<string, string> { { NamespaceKey, _namespaceId.ToString() } };
-            await container.CreateIfNotExistsAsync(metadata: metadata, cancellationToken: cancellationToken);
+            await _blobContainer.CreateIfNotExistsAsync(metadata: metadata, cancellationToken: cancellationToken);
 
             _logger.LogDebug("Fetching blob reference with name {ObjectName}", path);
             try
             {
 
-                await container.GetBlobClient(path).UploadAsync(content, cancellationToken);
+                await _blobContainer.GetBlobClient(path).UploadAsync(content, cancellationToken);
             }
             catch (RequestFailedException e)
             {
@@ -184,7 +174,7 @@ namespace Jupiter.Implementation
             {
                 // we touch the blob so that the last access time is always refreshed even if we didnt actually mutate it to make sure the gc knows this is a active blob
                 // see delete operation in Leda blob store cleanup
-                await TouchBlob(container.GetBlobClient(path));
+                await TouchBlob(_blobContainer.GetBlobClient(path));
                 _logger.LogDebug("Upload of blob {ObjectName} completed", path);
             }
         }
@@ -203,15 +193,15 @@ namespace Jupiter.Implementation
 
         public async Task<BlobContents?> TryReadAsync(string path, LastAccessTrackingFlags flags, CancellationToken cancellationToken)
         {
-            BlobContainerClient container = new BlobContainerClient(_connectionString, _containerName);
-            if (!await container.ExistsAsync(cancellationToken))
+
+            if (!await _blobContainer.ExistsAsync(cancellationToken))
             {
-                throw new InvalidOperationException($"Container {_containerName} did not exist");
+                throw new InvalidOperationException($"Container {_blobContainer.Name} did not exist");
             }
 
             try
             {
-                BlobClient blob = container.GetBlobClient(path);
+                BlobClient blob = _blobContainer.GetBlobClient(path);
                 Response<BlobDownloadInfo> blobInfo = await blob.DownloadAsync(cancellationToken);
                 return new BlobContents(blobInfo.Value.Content, blobInfo.Value.ContentLength);
             }
@@ -228,37 +218,34 @@ namespace Jupiter.Implementation
 
         public async Task<bool> ExistsAsync(string path, CancellationToken cancellationToken)
         {
-            BlobContainerClient container = new BlobContainerClient(_connectionString, _containerName);
-            if (!await container.ExistsAsync(cancellationToken))
+            if (!await _blobContainer.ExistsAsync(cancellationToken))
             {
                 return false;
             }
 
-            BlobClient blob = container.GetBlobClient(path);
+            BlobClient blob = _blobContainer.GetBlobClient(path);
             return await blob.ExistsAsync(cancellationToken);
         }
 
         public async Task DeleteAsync(string path, CancellationToken cancellationToken)
         {
-            BlobContainerClient container = new BlobContainerClient(_connectionString, _containerName);
-            if (!await container.ExistsAsync(cancellationToken))
+            if (!await _blobContainer.ExistsAsync(cancellationToken))
             {
-                throw new InvalidOperationException($"Container {_containerName} did not exist");
+                throw new InvalidOperationException($"Container {_blobContainer.Name} did not exist");
             }
 
-            await container.DeleteBlobAsync(path, cancellationToken: cancellationToken);
+            await _blobContainer.DeleteBlobAsync(path, cancellationToken: cancellationToken);
         }
 
         public async IAsyncEnumerable<(string, DateTime)> ListAsync([EnumeratorCancellation] CancellationToken cancellationToken)
         {
-            BlobContainerClient container = new BlobContainerClient(_connectionString, _containerName);
-            bool exists = await container.ExistsAsync(cancellationToken);
+            bool exists = await _blobContainer.ExistsAsync(cancellationToken);
             if (!exists)
             {
                 yield break;
             }
 
-            await foreach (BlobItem? item in container.GetBlobsAsync(BlobTraits.Metadata, cancellationToken: cancellationToken))
+            await foreach (BlobItem? item in _blobContainer.GetBlobsAsync(BlobTraits.Metadata, cancellationToken: cancellationToken))
             {
                 yield return (item.Name, item.Properties?.LastModified?.DateTime ?? DateTime.Now);
             }